US25 and VSCode Config
I added a field childrenIds to the Family class. This now stores the Ids uniquely of the children and the field children now stores the entire individual. The family also now has a method unique_family_names that can be used to see if each member of the family has a unique first name.